requirements.txt
bs4
openai
python-dotenv
numpy
transformers